Per-Op Medicine Prescribing Guidance
Began life at the #Gashack event at the Royal Colloege of Anaesthetists in London in October 2017. The aim is to create a repository of prescribing instructions for commonly-used medicines when the patient is fasting for surgery.
It may be as simple as "withhold" or as complex as peri-op insulin infusions e.g. DKI regimens.